Added deleting mods, changing priority, creating new priority
The mod table now has all the same functionality the original desktop app version does (except for displaying priority colors). The X button now properly removes mods from the table on both the front and back end. Same goes for the priority dropdown, which also has an additional option to create a new priority level. This reveals a popup that allows the user to enter a name and select a color for their new priority level, which is then added to the back end and displayed on the front end. Some other optimizations were made to support these changes, such as the use of memos for modList and priorityList in App.jsx