Version input, reload button, and table colors
The user now has a space where they can enter a version for their profile. Clicking the new reload button will update the selected version, as well as refresh the data with new API calls. Additionally, the color of each priority is shown in the 3rd column and the dropdown menu. If a mod is ready for the selected version, the version column will be green and have a ✔