Reintroduced download ready mods button
The download ready mods button has been reintroduced, as well as some minor refactoring to make it work. The feature works the same way as in the desktop app (except the tabs open faster and automatically close!). In mod.downloadMod() instead of returning True when a download link is found, the download link is returned. File download tests should no longer pass if the download link is None

Issue discovered while building this feature: CurseForge is returning None for downloadUrl on a lot of mods. Not sure what to do about this right now. This issue impacts previous versions of Mod Tracker and is not a problem with the app itself.