Support for multiple profiles, export button
Multiple profiles can now be read from local storage and accessed from localhost:5173/profile=[profileIndex]. To support this, the data stored in the browser's internal storage will now be the contents of the ProfileManager, rather than just storing a single profile.

main.jsx was also updated to show the new App component, which owns the ProfileFetcher (now DataFetcher) and decides whether to show ProfileSelectWindow or DetailsWindow. ProfileSelectWindow will be shown if no profile=[index] parameter is appended to the URL, or if the index is out of range. ProfileSelectWindow is just a skeleton right now.

A new export button has been added to the DetailsWindow, which downloads a json file structured the same way as the desktop app's export feature.

pyServer.py and appData.py were updated with three new calls: /restore-data, /get-num-profiles, and /restore-profile. Only /restore-data is actually used.