pyServer now supports multiple connected users
pyServer and its components have been significantly refactored in order to support multiple users connecting to the site at the same time. There is no longer a single DataManager that pyServer uses. Instead, pyServer creates a new DataManager for every call, loads all of the current data into it, carries out the desired operation, then returns the a new copy of the data with the operation complete. In short, pyServer no longer stores a DataManager, but takes the frontend's data, operates on it with a temporary DataManager, and returns new data. Now, mulitple users on the site at once won't cause conflicts, because they won't be fighting over the same DataManager object.

Also, fixed a crash when the localstorage is set to undefined.