Loading text, more tests, CSS tweaks, bugfixes
The function output text has been added to the profile select screen, and will now display "loading..." when a call to the backend is being made. More tests for restoreData were added to test_pyServer.py. More tweaks were made to the CSS, especially to the mod table. Only one scroll bar appears in the details window now, and it the user doesn't need to scroll to see the add mod input and button. A new profile is no longer automatically created when there is exactly one profile in the profile list.