Added 404.html, new URL structure
Vite and PyServer are currently working. However, there was a bug where if a profile was opened, bringing the user to /profile=0, it would not only incorrectly update the URL, but this method is unsupported by GitHub Pages anyway. As a workaround, a 404.html file as been added which mirrors index.html. Additionally, the URL structure has been changed from localhost/profile=0 to localhost/ModTracker/profile/0 to work better with GitHub Pages.