Refactor to single page display
Instead of appending the URL to navigate between screens, a simpler single page approach is now used by setting profileIndex (now curProfileIndex) manually. This system is faster and compatible with GitHub Pages, but unfortunately, does not allow users navigate directly to the details view of a profile by saving the extended URL. This is a necessary trade-off to get the site working on GitHub Pages.