pyServer_IP secret, bugfixed first site load
There is now a standalone pyserver hosted somewhere in the world. When hosted on GitHub Pages, the IP, which is held in a GitHub Action Secret, is loaded into pyServer_IP.txt, which is then read by dataFetcher. If the hostname is not GitHub Pages, dataFetcher defaults to the old method.

Addtionally, some bugs that occur when a user logs into the site for the first time (and thus has empty browser storage) have been addressed. A new profile is automatically created ONLY if the user does not have any profiles in their profile list, or the browser data is empty. Additionally, numProfiles in App.jsx now updates when the profileData does, so the profile tiles will always be clickable without reloading the page.