Fixed workflow, dataFetcher uses dynamic import
workflow should no longer output just "${API_Key}" but instead fill it in with the actual secret. Additionally, dataFetcher no longer imports pyServer_IP.txt from the top, but instead imports it dynamically with the import() function. This should allow the site to compile, regardless of if the site exists or not.