New method of injecting pyServer IP
Workflow has been updated to instead of creating a txt file, a config.js file is modified, and the dataFetcher creates the URL using that. This system should be a lot more resilient and applicable to GitHub Pages, as well as offering a way to easily change the pyServer URL in a development environment.