Folders renamed, deprecated files removed, new files created
For consistency, all the new folders in WebApp are now lowercase. Additionally, new files that will be used to rewrite the backend in JS have been added to the new WebApp/data folder (these are subject to change though). Finally, the deprecated PyServer and data fetcher files have been removed.