ProfileManager tests
The ProfileManager now has a full suite of tests for all of its public functions, leaving loadFromJson as the last backend module that needs to be rewritten and tested. ProfileManager is also now the sole source of the priority list, with Profile objects no longer having their own priority list.