Frontend adapted to render new backend
The frontend now reads a profile manager stored in stateProvider rather than calling a separate backend server. To accomplish this, lost of refactoring was required. To ensure react rerenders when changes are made to profileManager, the valtio libary is being used to manage state (although a force rerender is sometimes still required, so I might ditch it in favor of a normal react context provider). Private variables have also been made technically public but denoted with an underscore so their state can be tracked by valtio. An error boundary was also added to main.jsx